restructure torchtitan files -- continued
