+++ .gitignore
@@ -13,6 +13,7 @@ out
 wandb
 
 torchtitan/datasets/**/*.model
+assets/**/*.model
 
 # temp files
 *.log
