--- torchtitan/datasets/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from torchtitan.datasets.hf_datasets import build_hf_dataloader
-
-__all__ = ["build_hf_dataloader"]
