+++ torchtitan/models/llama/train_configs/llama3_70b.toml
@@ -19,7 +19,7 @@ save_tb_folder = "tb"
 name = "llama3"
 flavor = "70B"
 norm_type = "rmsnorm"  # layernorm / np_layernorm / rmsnorm
-tokenizer_path = "./torchtitan/datasets/tokenizer/original/tokenizer.model"
+tokenizer_path = "./assets/tokenizer/original/tokenizer.model"
 # converters = "float8"
 
 [optimizer]
